chore: divide landing page into server and client components for metadata
